requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
cachetools==5.3.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
control==0.9.4
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.103.1
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.16.3
flatbuffers==23.3.3
flower==2.0.1
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
kombu==5.3.2
libclang==16.0.0
magic-filter==1.0.12
makefun==1.15.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
metakernel==0.30.1
mistune==2.0.5
ml-dtypes==0.0.4
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.5
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pygame==2.5.2
Pygments==2.15.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
